feat(Frontend): :art: Move wallet to navbar group and change date format function
